refactor: 내부 구조 변경
