+++ pull-request-template.md
@@ -1,0 +1,8 @@
+## 반영 브랜치
+profile -> develop
+
+## 변경 사항
+기초 작업 및 연습을 진행했습니다.
+
+## 테스트 결과
+(아직 없음)
